backend/requirements.txt
django>=4.0
djangorestframework
django-cors-headers
psycopg2-binary
celery
redis
pandas
openpyxl
